lib: cobs: implement streaming cobs variant
Many transports deliver data in small, arbitrary chunks (for example,
UART RX buffers). In these cases, treating COBS encoding/decoding as a
single “whole buffer in / whole buffer out” operation is awkward.

This paves the way for being able to add highly efficient cobs-based
interfaces that operate on chunks instead of whole buffers. We avoid
creating a new buffer just to hold incomplete frame and instead maintain
state as we are receiving more data.

Encoding is a bit more complex since cobs requires lookahead when
encoding, but this is not a problem because we have access to full data
when we encode so we can do lookahead when needed.